Major overhaul with the underlying JavaScript
A lot of compatability has been added. It can now handle keynames with special characters in it. It can now handle keynames that start with a number.

The Tree View to JSON code is around 25 to 50 times faster.

The JSON to Tree View code is about half as fast, but improvements may be made to make this faster.

The code is cleaner than before.