Fixed a bug that happened rarely when you try to delete an element.
Sometimes when you delete a backboneElement and if the element has a multiple cardinality and it was the last sub-element, and you clicked the delete button then the function would crash and the element wouldn't be deleted. This was fixed by adding a try catch and having an additional conditional check to make sure an element exists before trying to give it the "+" button.